[style] #90 파일 분리
