+++ presentation/src/main/java/com/mashup/presentation/chat/compose/ChatListScreen.kt
@@ -1,0 +1,13 @@
+package com.mashup.presentation.chat.compose
+
+import androidx.compose.runtime.Composable
+
+/**
+ * Ssam_D_Android
+ * @author jaesung
+ * @created 2023/06/27
+ */
+@Composable
+fun ChatListScreen() {
+
+}
